Refactor PhoSim to reduce duplication with ImSim
